Modificacion vista de configRed
