+++ static/css/model.css
@@ -1,114 +1,151 @@
-    .containerSlider {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        text-align: center;
-        font-weight: bold;
-    }
+.containerSlider {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    font-weight: bold;
+}
 
-    .info {
-        text-align: center;
-        color: #fff;
-        margin-bottom: 20px;
-    }
+.container-configRed {
+    margin-bottom: 40px;
+}
 
-    .range-slider {
-        margin-top: 0vh;
-    }
+.training-param {
 
-    .rs-label {
-        position: relative;
-        transform-origin: center center;
-        display: block;
-        width: 38px;
-        height: 38px;
-        background: transparent;
-        border-radius: 50%;
-        line-height: 30px;
-        text-align: center;
-        font-weight: bold;
-        padding-top: 6px;
-        box-sizing: border-box;
-        border: 2px solid #fff;
-        margin-top: 0px;
-        margin-left: -6px;
-        left: attr(value);
-        color: #fff;
-        font-style: normal;
-        font-weight: normal;
-        line-height: normal;
-        font-size: 22px;
-    }
+}
 
-    .rs-range {
-        margin-top: 15px;
-        width: 300px;
-        -webkit-appearance: none;
-    }
-    .rs-range:focus {
-        outline: none;
-    }
-    .rs-range::-webkit-slider-runnable-track {
-        width: 100%;
-        height: 1px;
-        cursor: pointer;
-        box-shadow: none;
-        background: #ffffff;
-        border-radius: 0px;
-        border: 0px solid #010101;
-    }
-    .rs-range::-moz-range-track {
-        width: 100%;
-        height: 1px;
-        cursor: pointer;
-        box-shadow: none;
-        background: #ffffff;
-        border-radius: 0px;
-        border: 0px solid #010101;
-    }
-    .rs-range::-webkit-slider-thumb {
-        box-shadow: none;
-        border: 0px solid #ffffff;
-        box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.25);
-        height: 22px;
-        width: 23px;
-        border-radius: 22px;
-        background: rgb(255, 255, 255);
-        cursor: pointer;
-        -webkit-appearance: none;
-        margin-top: -10px;
-    }
-    .rs-range::-moz-range-thumb {
-        box-shadow: none;
-        border: 0px solid #ffffff;
-        box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.25);
-        height: 22px;
-        width: 22px;
-        border-radius: 22px;
-        background: rgb(255, 255, 255);
-        cursor: pointer;
-        -webkit-appearance: none;
-        margin-top: -20px;
-    }
-    .rs-range::-moz-focus-outer {
-        border: 0;
-    }
-    
-    .box-minmax {
-        margin-top: 10px;
-        width: 308px;
-        display: flex;
-        justify-content: space-between;
-        font-size: 20px;
-        color: #FFFFFF;
-    }
-    .box-minmax span:first-child {
-        margin-left: 10px;
-    }
+.info {
+    text-align: center;
+    color: #fff;
+    margin-bottom: 20px;
+}
 
-    .container-sliderNeuronas {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
+.range-slider {
+    margin-top: 0vh;
+}
+
+.rs-label {
+    position: relative;
+    transform-origin: center center;
+    display: block;
+    width: 38px;
+    height: 38px;
+    background: transparent;
+    border-radius: 50%;
+    line-height: 30px;
+    text-align: center;
+    font-weight: bold;
+    padding-top: 6px;
+    box-sizing: border-box;
+    border: 2px solid #fff;
+    margin-top: 0px;
+    margin-left: -6px;
+    left: attr(value);
+    color: #fff;
+    font-style: normal;
+    font-weight: normal;
+    line-height: normal;
+    font-size: 22px;
+}
+
+.rs-range {
+    margin-top: 15px;
+    width: 300px;
+    -webkit-appearance: none;
+}
+.rs-range:focus {
+    outline: none;
+}
+.rs-range::-webkit-slider-runnable-track {
+    width: 100%;
+    height: 3px;
+    cursor: pointer;
+    box-shadow: none;
+    background: #ffffff;
+    border-radius: 0px;
+    border: 0px solid #010101;
+}
+.rs-range::-moz-range-track {
+    width: 100%;
+    height: 1px;
+    cursor: pointer;
+    box-shadow: none;
+    background: #ffffff;
+    border-radius: 0px;
+    border: 0px solid #010101;
+}
+.rs-range::-webkit-slider-thumb {
+    box-shadow: none;
+    border: 0px solid #ffffff;
+    box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.25);
+    height: 22px;
+    width: 23px;
+    border-radius: 22px;
+    background: #28a328;
+    cursor: pointer;
+    -webkit-appearance: none;
+    margin-top: -10px;
+}
+.rs-range::-moz-range-thumb {
+    box-shadow: none;
+    border: 0px solid #ffffff;
+    box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.25);
+    height: 22px;
+    width: 22px;
+    border-radius: 22px;
+    background: rgb(255, 255, 255);
+    cursor: pointer;
+    -webkit-appearance: none;
+    margin-top: -20px;
+}
+.rs-range::-moz-focus-outer {
+    border: 0;
+}
+
+.box-minmax {
+    margin-top: 10px;
+    width: 308px;
+    display: flex;
+    justify-content: space-between;
+    font-size: 20px;
+    color: #FFFFFF;
+}
+.box-minmax span:first-child {
+    margin-left: 10px;
+}
+
+.container-sliderNeuronas {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 10px;
+}
+
+.activation-function-select {
+    /* General appearance */
+    display: inline-block;
+    padding: 5px 10px;
+    border: 1px solid #663300; /* Dark mahogany border */
+    border-radius: 5px;
+    background-color: #6d4423; /* Dark mahogany background */
+    color: #fff; /* White text */
+    font-family: sans-serif; /* Choose a non-modern sans-serif font */
+}
+.activation-function-select:hover {
+    background-color: #805029; /* Slightly darker mahogany on hover */
+}
+.activation-function-select:focus {
+    outline: none; /* Remove default outline */
+    box-shadow: 0 0 5px #805029; /* Dark mahogany inner shadow */
+}
+.activation-function-select:disabled {
+    opacity: 0.5; /* Reduced opacity for disabled elements */
+    cursor: not-allowed; /* Indicate disabled state */
+}
+.activation-function-select option {
+    background-color: #6d4423; /* Dark mahogany background */
+    color: #fff; /* White text */
+    padding: 5px;
+    cursor: pointer;
+}
