--- static/scripts/utils.js
@@ -1,3 +1,0 @@
-const utils = {
-}
-export default utils;
